+++ src/main/resources/static/css/order.css
@@ -6,6 +6,76 @@ body {
     flex-direction: column;
 }
 
+.order-table {
+    width: 100%;
+    margin: 20px auto;
+    border-collapse: collapse;
+    background-color: #fff;
+}
+
+.order-table thead tr {
+    border-top: 2px solid #a7a7a7;
+    border-bottom: 2px solid #a7a7a7;
+}
+
+.order-table th, .order-table td {
+    padding: 15px;
+    text-align: center;
+    border-bottom: 1px solid #e9e9e9;
+}
+
+.order-table th {
+    background-color: #f4f4f4;
+    font-weight: bold;
+}
+
+.product-info {
+    display: flex;
+    align-items: center;
+}
+
+.product-image {
+    width: 50px;
+    height: 50px;
+    margin-right: 15px;
+}
+
+.product-details {
+    text-align: left;
+}
+
+.buy-button, .delete-button {
+    padding: 10px 20px;
+    margin: 5px;
+    font-size: 14px;
+    color: #fff;
+    background-color: #000;
+    border: none;
+    cursor: pointer;
+}
+
+.buy-button {
+    background-color: #007bff;
+}
+
+.delete-button {
+    background-color: #4fbbe8;
+}
+
+.buy-button:hover, .delete-button:hover {
+    opacity: 0.8;
+}
+
+@media (max-width: 768px) {
+    .order-table, .product-info {
+        display: block;
+    }
+
+    .product-details {
+        margin-top: 10px;
+    }
+}
+
 .order-container {
     margin: 30px 50px;
     background-color: white;
@@ -124,7 +194,7 @@ form .form-group select {
 }
 
 .order-btn {
-    background-color: #333;
+    background-color: #4fbbe8;
     color: white;
 }
