+++ src/main/resources/templates/echopickup/product/order.html
@@ -13,30 +13,37 @@
 <div layout:fragment="content">
   <main>
     <div class="order-container">
-      <h2>배송지 정보</h2>
+      <table id="orderItemsTable" class="order-table">
+        <thead>
+        <tr>
+          <th>제품정보</th>
+          <th>결제 금액</th>
+          <th>선택</th>
+        </tr>
+        </thead>
+        <tbody id="classifiedItems">
+        <!-- 분류된 항목들이 여기 추가됩니다 -->
+        </tbody>
+      </table>
+
+      <h2>수거 정보</h2>
       <form action="#" method="post">
         <div class="form-group">
-          <label for="name">받는 분 *</label>
+          <label for="name">신청인</label>
           <input type="text" id="name" name="name" placeholder="이름을 입력하세요">
         </div>
         <div class="form-group">
-          <label for="email">이메일 *</label>
-          <input type="email" id="email" name="email" placeholder="이메일을 입력하세요">
-        </div>
-        <div class="form-group">
-          <label for="phone1">휴대전화 *</label>
+          <label for="phone">휴대전화</label>
           <div class="phone-input">
-            <input type="tel" id="phone1" name="phone1" placeholder="010">
-            <input type="tel" id="phone2" name="phone2" placeholder="1234">
-            <input type="tel" id="phone3" name="phone3" placeholder="5678">
+            <input type="tel" id="phone" name="phone" placeholder="01012345678">
           </div>
         </div>
         <div class="form-group">
-          <label for="date">수거 예정일 *</label>
+          <label for="date">수거 예정일</label>
           <input type="date" id="date" name="date">
         </div>
         <div class="form-group">
-          <label for="address">주소 *</label>
+          <label for="address">주소</label>
           <input type="text" id="address" name="address" placeholder="주소를 입력하세요">
           <button type="button" class="find-address-btn">우편번호</button>
         </div>
@@ -58,7 +65,6 @@
       <div class="payment-method">
         <label><input type="radio" name="payment" value="credit" checked> 신용카드</label>
         <label><input type="radio" name="payment" value="kakaopay"> 카카오페이</label>
-        <label><input type="radio" name="payment" value="paypal"> 페이팔</label>
         <label><input type="radio" name="payment" value="naverpay"> 네이버페이</label>
         <label><input type="radio" name="payment" value="bank"> 무통장 입금</label>
       </div>
@@ -81,6 +87,8 @@
         <button type="submit" class="order-btn">결제하기</button>
       </div>
     </div>
+
+    <script th:src="@{/js/order.js}"></script>
   </main>
 </div>
 </body>
