+++ src/main/resources/templates/echopickup/product/product.html
@@ -42,8 +42,9 @@
         <table id="classifiedItemsTable" border="1" style="width: 100%; border-collapse: collapse;">
             <thead>
             <tr>
-                <th style="width: 50%;">제품명</th>
-                <th style="width: 50%;">가격</th>
+                <th style="width: 30%;">사진</th>
+                <th style="width: 30%;">제품명</th>
+                <th style="width: 30%;">가격</th>
             </tr>
             </thead>
             <tbody id="classifiedItems">
@@ -72,8 +73,8 @@
             </div>
         </div>
 
+        <script th:src="@{/js/product.js}"></script>
     </main>
 </div>
-
 </body>
 </html>
